+++ Project/WorldOfFacades/TexturedObjectSlide.h
@@ -37,10 +37,10 @@ public:
 	CTexturedObjectSlide(SVector2 _pos, SVector2 _size) : CTexturedObject(_pos)
 	{
 		m_position = _pos;
-		m_rect.x = _pos.X;
-		m_rect.y = _pos.Y;
-		m_rect.w = _size.X;
-		m_rect.h = _size.Y;
+		m_rect.x = (int)_pos.X;
+		m_rect.y = (int)_pos.Y;
+		m_rect.w = (int)_size.X;
+		m_rect.h = (int)_size.Y;
 	}
 
 	/// <summary>
